backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Payments
stripe==10.12.0

# Media / Storage
pillow==10.4.0
python-multipart==0.0.12

# Async tasks
redis==5.1.1

# Utils
python-dateutil==2.9.0.post0
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
